requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

discord.py==2.3.1
python-dotenv==0.20.0
